Allow sqlite3 to create cursors on-the-fly
To avoid having a cursor object around when it is not needed, simply use
the non-standard shortcut methods on the database object to create a
cursor for the length of the SQL statement(s) execution. This way the
cursor will be cleaned up immediately after the SQL transaction has
completed.